# docs/swagger/CSaS-swagger.json
{"openapi":"3.0.0","info":{"title":"Estudio de Caso como Servicio (CSaaS)","description":"","version":"0.1","contact":{}},"tags":[],"servers":[],"components":{"schemas":{"EstudioCaso":{"type":"object","properties":{"id":{"type":"number"},"nombre":{"type":"string"},"intervenciones":{"type":"array","items":{"type":"string"}}},"required":["id","nombre","intervenciones"]},"CreateEstudioCasoDto":{"type":"object","properties":{}},"UpdateEstudioCasoDto":{"type":"object","properties":{}},"Caso":{"type":"object","properties":{"id":{"type":"number"},"nombre":{"type":"string"},"descripcion":{"type":"string"},"preguntas":{"type":"array","items":{"type":"string"}},"proposicionesTeoricas":{"type":"array","items":{"type":"string"}},"unidadesAnalisis":{"type":"array","items":{"type":"string"}}},"required":["id","nombre","descripcion","preguntas","proposicionesTeoricas","unidadesAnalisis"]},"CreateCasoDto":{"type":"object","properties":{"nombre":{"type":"string"},"descripcion":{"type":"string"},"idEstudioCaso":{"type":"number"},"preguntas":{"type":"array","items":{"type":"string"}}},"required":["nombre","descripcion","idEstudioCaso","preguntas"]},"CreateProposicionTeoricaDto":{"type":"object","properties":{"nombre":{"type":"string"}},"required":["nombre"]},"ProposicionToCase":{"type":"object","properties":{}},"UnidadAnalisis":{"type":"object","properties":{"id":{"type":"number"},"nombre":{"type":"string"},"descripcion":{"type":"string"},"tipo":{"type":"string"},"interacciones":{"type":"array","items":{"type":"string"}}},"required":["id","nombre","descripcion","tipo","interacciones"]},"CreateUnidadAnalisisDto":{"type":"object","properties":{"nombre":{"type":"string"},"descripcion":{"type":"string"},"tipo":{"type":"string"},"idCaso":{"type":"number"},"componentes":{"type":"array","items":{"type":"string"}}},"required":["nombre","descripcion","tipo","idCaso","componentes"]},"UpdateUnidadAnalisisDto":{"type":"object","properties":{"nombre":{"type":"string"},"descripcion":{"type":"string"},"tipo":{"type":"string"},"idCaso":{"type":"number"},"componentes":{"type":"array","items":{"type":"string"}}},"required":["nombre","descripcion","tipo","idCaso","componentes"]},"CreateInteraccionDto":{"type":"object","properties":{"idUnidadAnalisis":{"type":"number"},"evidencias":{"type":"array","items":{"type":"string"}}},"required":["idUnidadAnalisis","evidencias"]}}},"paths":{"/estudioCaso":{"get":{"operationId":"EstudioCasoController_getAll","parameters":[{"name":"idConsultoria","required":true,"in":"query","schema":{"type":"number"}}],"responses":{"200":{"description":"Pedir Estudios de Caso por Consultoria","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/EstudioCaso"}}}}}},"tags":["Estudio de Caso"]},"post":{"operationId":"EstudioCasoController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateEstudioCasoDto"}}}},"responses":{"201":{"description":"Insertar Estudio de Caso","content":{"application/json":{"schema":{"$ref":"#/components/schemas/EstudioCaso"}}}}},"tags":["Estudio de Caso"]}},"/estudioCaso/{id}":{"get":{"operationId":"EstudioCasoController_getOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":"Pedir Estudios de Caso por Id","content":{"application/json":{"schema":{"$ref":"#/components/schemas/EstudioCaso"}}}}},"tags":["Estudio de Caso"]},"put":{"operationId":"EstudioCasoController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateEstudioCasoDto"}}}},"responses":{"200":{"description":""}},"tags":["Estudio de Caso"]}},"/caso":{"get":{"operationId":"CasoController_getAllByEstudioCaso","parameters":[{"name":"id","required":true,"in":"query","schema":{"type":"number"}}],"responses":{"200":{"description":"Pedir Casos por Estudio de Caso","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Caso"}}}}}},"tags":["Caso"]},"post":{"operationId":"CasoController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateCasoDto"}}}},"responses":{"201":{"description":"Insertar Estudio de Caso","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Caso"}}}}},"tags":["Caso"]}},"/caso/{id}":{"get":{"operationId":"CasoController_getOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":"Pedir Caso por Id","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Caso"}}}}},"tags":["Caso"]},"put":{"operationId":"CasoController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["Caso"]}},"/caso/proposicionTeorica":{"post":{"operationId":"CasoController_addProposicionTeorica","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateProposicionTeoricaDto"}}}},"responses":{"201":{"description":""}},"tags":["Caso"]}},"/caso/proposicionTeorica/{id}":{"delete":{"operationId":"CasoController_removeProposicionTeorica","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["Caso"]}},"/caso/addProposicionToCaso":{"post":{"operationId":"CasoController_addToCase","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateCasoDto"}}}},"responses":{"201":{"description":"Insertar Estudio de Caso"}},"tags":["Caso"]}},"/caso/removeProposicionFromCaso":{"post":{"operationId":"CasoController_removeFromCase","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProposicionToCase"}}}},"responses":{"201":{"description":""}},"tags":["Caso"]}},"/unidadAnalisis":{"get":{"operationId":"UnidadAnalisisController_findAll","parameters":[{"name":"idCaso","required":true,"in":"query","schema":{"type":"number"}}],"responses":{"200":{"description":"Listar unidades","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/UnidadAnalisis"}}}}}},"tags":["Unidad de Análisis"]},"post":{"operationId":"UnidadAnalisisController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUnidadAnalisisDto"}}}},"responses":{"201":{"description":"Creada","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUnidadAnalisisDto"}}}}},"tags":["Unidad de Análisis"]}},"/unidadAnalisis/{id}":{"get":{"operationId":"UnidadAnalisisController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":"Pedir Unidad por su Id","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UnidadAnalisis"}}}}},"tags":["Unidad de Análisis"]},"put":{"operationId":"UnidadAnalisisController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateUnidadAnalisisDto"}}}},"responses":{"201":{"description":"Creada","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UnidadAnalisis"}}}}},"tags":["Unidad de Análisis"]}},"/unidadAnalisis/nomencladores/tiposComponentes":{"get":{"operationId":"UnidadAnalisisController_getTiposComponente","parameters":[],"responses":{"201":{"description":"Creada","content":{"application/json":{"schema":{"type":"array","items":{"type":"string"}}}}}},"tags":["Unidad de Análisis"]}},"/unidadAnalisis/interaccion":{"post":{"operationId":"UnidadAnalisisController_addInteraccion","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateInteraccionDto"}}}},"responses":{"201":{"description":""}},"tags":["Unidad de Análisis"]}}}}