# src/modules/caso/dto/create-caso.dto.ts
import { ApiProperty } from "@nestjs/swagger";

export class CreateCasoDto {
    @ApiProperty()
    nombre: string;

    @ApiProperty()
    descripcion: string;

    @ApiProperty()
    idEstudioCaso: number;

    @ApiProperty()
    preguntas: string[];
}