# src/modules/unidad-analisis/dto/create-interaccion.dto.ts
import { EvidenciaDto } from "./evidencia.dto";
import { ApiProperty } from "@nestjs/swagger";

export class CreateInteraccionDto {
    @ApiProperty()
    idUnidadAnalisis: number;

    @ApiProperty()
    evidencias: EvidenciaDto[];
}